move AllocProfile into Profile package
