--- stdlib/AllocProfile/Project.toml
@@ -1,2 +1,0 @@
-name = "AllocProfile"
-uuid = "0009037c-06ea-4b9e-a52e-4b728900de75"
